Updates managers.js
Separates and includes functions to toggle the view from cards to rows in both the home page and the specific intern profile. It also modifies the report and submission functions to match is view. In addition, since now each card and submission row includes approved by, the /report, /approvebymanager, /manageredir and /edit rows also support this new feature. Moreover, the edit functions now erase the information about the previous approval once it is approved. In other words, once the manager decides to edit one submission, it erases the approved by name and date of approval from the submission. Last, the feature approved by was also added to the /approveall route.