Updates mongo_with_gsheets.js
Adds conditional to check whether it is the first week of the pay period and run the consolidation of the databases once every two weeks. This reduces the number of requests to mongodb Realm and also permits a buffer time for employees and interns to have a last pay period available to be paid on all the pending salaries, if any.