Updates managers.ejs
Modifies navbar collapse and the options to most recent route. Adds Toggle View from cards to row view option on the navbar. Modifies the structure and view of the cards. Includes the option to show the data as rows. Modifies what goes inside the cards, more information taking a better advantage of the space.