Updates profile.ejs
Includes ToggleView button. Modifies the structure of the cards to take a better advantage of the space and avoid items flying out of the cards visual. Includes the view to render both cards and rows. More information is now encompassed in both the cards and the rows. Options to edit and approve remain.