Updates profile.ejs
Updates navbar logo and color. Adds support for Active/Inactive, Average and Total badges and popovers according the numbers in the submission and personal employee stats. Modifies wording from Missed to Unreported and Justification/Tasks Developed to Tasks Completed. Adds Popovers to explain the meaning of Unreported and Total. Adds these features previously mentioned to both row and cards view. Modifies UI to reformat the cards. Debuggs card items that were flying out of it and their arrangement. Modified the UI to make distinctions on the sizes and arrangements of the items. Increases reference to pay periods dashboards from interns rows. Increases the size of textareas on cards.