Creates gsheetAPIs.js
This file contains many functions necessary to bridge the app with the main Google Sheet for management using Google Sheets API. The async functions drag data from a specific sheet according to the parameters on the keys file. It then reports this data to other important elements of the app.