Updates dashboard.ejs
Adds partials for header and footer. Adds more elements of the profile stripe. Modifies the layout of the cards. Modifies the wording of the popovers. Modifies the sizes and items inside the cards. Increases the width of the table in the UI. Eliminates the weird white gap between the first line of cards and the second one.