Updates managers.ejs
Adds headers and footers as partials. Modifies the layout of the cards. Modifies the UI of the images. Increased the width of the table to benefit the textarea with the tasks completed. Moved the approve button to mostly the right. Eliminated email as an element in the row view to benefit the size of the tasks completed textarea.